services/auth/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-service"
version = "0.1.0"
description = "Production-ready auth microservice"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "redis>=5.0.8,<6.0.0",
  "pyjwt[crypto]>=2.9.0,<3.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "structlog>=24.4.0,<25.0.0",
  "nats-py>=2.8.0,<3.0.0",
  "arq>=0.25.0,<1.0.0",
  "opentelemetry-api>=1.27.0,<2.0.0",
  "opentelemetry-sdk>=1.27.0,<2.0.0",
  "opentelemetry-exporter-otlp>=1.27.0,<2.0.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0,<1.0.0",
  "opentelemetry-instrumentation-asyncpg>=0.48b0,<1.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "asgi-lifespan>=2.1.0,<3.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["auth_service*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["auth_service/tests"]
